+++ tests/test_tdom.py
@@ -1,16 +1,16 @@
 from typing import Sequence
 
-from tdom.tdom import (
+from tdom import (
     Element,
     Fragment,
     Text,
+    unsafe,
+)
+from tdom.tdom import (
     _as_node,
     _serialize_attributes,
     get_component_value,
-    unsafe,
 )
-
-# Tests for get_component_value
 
 
 def test_props_and_target_no_children():
@@ -52,31 +52,7 @@ def test_props_yes_target_yes():
     assert result == 3
 
 
-def test_micropython_no_children():
-    def target():
-        return 99
-
-    props = {}
-    _children = []
-    context = {}
-    imp = True  # Simulate running under MicroPython
-    result = get_component_value(props, target, _children, context, imp=imp)
-    assert result == 99
-
-
-def test_micropython_with_children():
-    def target(children: Sequence):
-        return len(children)
-
-    props = {}
-    _children = [1, 2, 3]
-    context = {}
-    imp = True  # Simulate running under MicroPython
-    result = get_component_value(props, target, _children, context, imp=imp)
-    assert result == 3
-
-
-def test_not_micropython_children_context_injection():
+def test_children_context_injection():
     """The target asks for both children and context."""
 
     def target(children, context):
